+++ plugin/src/main/java/com/skyblock/plugin/profile/SkyBlockProfile.java
@@ -25,7 +25,7 @@ public final class SkyBlockProfile implements Serializable {
     private static final long serialVersionUID = 1L;
 
     private final UUID uuid;
-    private final Map<String, Long> skillXp = new HashMap<>();
+    private final Map<String, Double> skillXp = new HashMap<>();
     private final Map<String, Long> collectionXp = new HashMap<>();
     private final Map<String, Integer> collectionCounts = new HashMap<>();
     private final List<ItemStack> quiverContents = new ArrayList<>();
@@ -64,7 +64,7 @@ public final class SkyBlockProfile implements Serializable {
      *
      * @return the skill experience totals
      */
-    public Map<String, Long> getSkillXp() {
+    public Map<String, Double> getSkillXp() {
         return Map.copyOf(skillXp);
     }
 
@@ -74,9 +74,9 @@ public final class SkyBlockProfile implements Serializable {
      * @param skill the skill name
      * @return the experience, or 0 if the skill has never been trained
      */
-    public long getSkillXp(String skill) {
+    public double getSkillXp(String skill) {
         Objects.requireNonNull(skill, "skill");
-        return skillXp.getOrDefault(skill, 0L);
+        return skillXp.getOrDefault(skill, 0.0);
     }
 
     /**
@@ -86,12 +86,12 @@ public final class SkyBlockProfile implements Serializable {
      * @param amount the experience to add, must not be negative
      * @throws IllegalArgumentException if {@code amount} is negative
      */
-    public void addSkillXp(String skill, long amount) {
+    public void addSkillXp(String skill, double amount) {
         Objects.requireNonNull(skill, "skill");
         if (amount < 0) {
             throw new IllegalArgumentException("amount must not be negative, got " + amount);
         }
-        skillXp.merge(skill, amount, Long::sum);
+        skillXp.merge(skill, amount, Double::sum);
     }
 
     /**
@@ -115,7 +115,7 @@ public final class SkyBlockProfile implements Serializable {
      * @param amount the new experience total, must not be negative
      * @throws IllegalArgumentException if {@code amount} is negative
      */
-    public void setSkillXp(String skill, long amount) {
+    public void setSkillXp(String skill, double amount) {
         Objects.requireNonNull(skill, "skill");
         if (amount < 0) {
             throw new IllegalArgumentException("amount must not be negative, got " + amount);
